Function demo_rcv::print_discovered_sources

fn print_discovered_sources(srcs: &Vec<DiscoveredSacnSource>)

Prints the given array of discovered sources to std out. Uses the format "Name: x, Universes: y" where x is the source name and y is the universes registered to the source.

Arguments

src: The sources to print to standard out.